Updated requested information
