+++ home/urls.py
@@ -7,5 +7,5 @@ app_name = "home"
 urlpatterns = [
     path("", views.home_view, name="home"),
     path("search", views.search_view, name="search"),
-    path("details", views.details_view, name="details"),
+    path("details/<str:id>/", views.details_view, name="details"),
 ]
