+++ pyproject.toml
@@ -9,7 +9,6 @@ readme = "README.md"
 python = "^3.11"
 django = "^5.0.1"
 pyyaml = "^6.0.1"
-htmx = "^0.0.0"
 gunicorn = "^21.2.0"
 whitenoise = "^6.6.0"
 ministryofjustice-data-platform-catalogue = "^0.7.0"
