+++ templates/base/head.html
@@ -11,6 +11,5 @@
   <link rel="manifest" href="{% static 'assets/manifest.json' %}">
   <link rel="stylesheet" type="text/css" href="{% static 'assets/css/search.css' %}">
   <link rel="stylesheet" type="text/css" href="{% static 'assets/css/base.css' %}">
-  <script src="https://unpkg.com/htmx.org@1.7.0/dist/htmx.js"></script>
   <meta property="og:image" content="{% static 'assets/images/govuk-opengraph-image.png' %}">
 </head>
